requirements.txt
transformers>=4.49.0
accelerate>=0.34.0
datasets>=2.21.0
trl>=0.14.0
peft>=0.14.0
qwen-vl-utils
wandb
typer
pyyaml
scipy
numpy
pillow
python-dotenv
